+++ src/app/components/token-editor.js
@@ -65,7 +65,7 @@ import PromiseObject from 'onedata-gui-common/utils/ember/promise-object';
 import PromiseArray from 'onedata-gui-common/utils/ember/promise-array';
 import safeExec from 'onedata-gui-common/utils/safe-method-execution';
 import computedT from 'onedata-gui-common/utils/computed-t';
-import RecordsOptionsArrayProxy from 'onezone-gui/utils/record-options-array-proxy';
+import RecordOptionsArrayProxy from 'onedata-gui-common/utils/record-options-array-proxy';
 
 const tokenInviteTypeOptions = [{
   value: 'userJoinGroup',
@@ -83,6 +83,11 @@ const tokenInviteTypeOptions = [{
   value: 'groupJoinSpace',
   icon: 'space',
   targetModelName: 'space',
+}, {
+  value: 'harvesterJoinSpace',
+  icon: 'space',
+  targetModelName: 'space',
+  noPrivileges: true,
 }, {
   value: 'userJoinCluster',
   icon: 'cluster',
@@ -1508,10 +1513,10 @@ export default Component.extend(I18n, {
   getRecordOptionsForModel(modelName) {
     return PromiseArray.create({
       promise: this.get('recordManager').getUserRecordList(modelName)
-        .then(records => get(records, 'list'))
-        .then(recordsList => RecordsOptionsArrayProxy.create({
+        .then(recordsList => get(recordsList, 'list'))
+        .then(records => RecordOptionsArrayProxy.create({
           ownerSource: this,
-          records: recordsList,
+          records,
         })),
     });
   },